Fix CI - Allow UINT32 in Generic Reduces (#31575)
### Ticket
NA

### Problem description
Validation was updated in
https://github.com/tenstorrent/tt-metal/commit/f8cf77360617b35bbdbb5adf570f6c5a3093ae32
for generic reduces to only accept floating 8/16/32 bit inputs, as the
kernel and underlying LLKs don't have full integer support.

It turns out the `VADv2` model is using the generic reduces with UNT32,
so this validation causes a failure in pipeilnes.

### What's changed
Update the validation immediately to allow UINT32 through, in order to
unblock this model.

Tests and documentation will be updated once support is better studied.

### Checklist
- [x] Frequent Model and TTNN Tests [VADv2
passes](https://github.com/tenstorrent/tt-metal/actions/runs/18972806289)