fix :bug:: overconstrained kt_dim in matmul tests (#774)
### Ticket
<!-- Link to Github Issue -->
#767 

### Problem description
It seems the the combination generator had overconstrained the `kt_dim`
because it assumed that the A and B matrices had to be small enough to
fit into Dest, when that is the case only for the resulting matrix

The Stochastic rounding exception is handled in `sweep_matmul`

### What's changed
<!-- Describe the approach used to solve the problem.
Summarize the changes made and its impact. -->

### Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update

### Checklist
<!-- These are required steps and need to be run from tt-metal
repository's Actions. Use links below and replace them with your run -->
- [ ] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes
- [ ] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml)
CI passes (if applicable)