Dataflow 2.0 API: Default constructor for Noc class added (#31711)
### Ticket
N/A

### Problem description
In a lot of noc use cases the noc id is implied to use the default
`noc_index`. In the current Noc class constructor in dataflow 2.0 APIs,
the noc id has to be passed in as an argument.

### What's changed
Added a default constructor for the Noc class that sets the internal
noc_id_ to the default `noc_index`, removing the need to pass in a noc
id argument when creating a Noc object.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/19046895258)
CI passes
- [x] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/19046898268)
CI tests passes