[skip ci]: ttnn: control executable check by --skip-executable-check flag (#31568)
### Ticket
https://github.com/tenstorrent/tt-metal/issues/31506

### Problem description
tt-run expect the remote program to be present on the launcher (host
that executes tt-run). This breaks the behavior of MPI Operator where
Launcher is just a tiny pod with MPI, and Workers are executing the
actual work and need to have dependencies.


### What's changed
New flag - `--skip-executable-check` (false by default) that skips the
check.