[TT-DiT] Add experimental fused distributed RMSNorm (#31689)
### Ticket
In service of https://github.com/tenstorrent/tt-metal/issues/29148

### Problem description
We have a common pattern of distributed RMSNorm -> create_heads -> ROPE
in the Wan2.2 model. This unfused sequence of operations accounts for
1.3% to 1.6% of total device time.
This sequence can be optimized with fusion. 

### What's changed
Introduce
- `ttnn.experimental.wan_fused_rmsnorm_pre_allgather`, an optimized,
streaming first stage of distributed RMSNorm
- `ttnn.experimental.wan_fused_rmsnorm_post_allgather`, an optimized,
streaming second stage of distributed RMSNorm with optional ROPE and
create_heads fusion

This operation is necessary on WH and BH galaxy for Wan2.2 performance.
The experimental namespace and the `wan` prefix should sufficiently warn
that this op is purpose-built.

The test was added to
`tests/ttnn/nightly/unit_tests/operations/transformers` such that it
should be ran in metal L2 nightly.

### Checklist
- [x] tt-metal l2 nightly
https://github.com/tenstorrent/tt-metal/actions/runs/19046711989

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>