Set default tensix extension mode to disabled in BH (#32008)
### Problem description
Currently tensix extension is enabled by default on BH, however, this
caused dispatch to be on row instead of col, and thus losing 2-4 cores
from compute grid. This causes some CCLs not fit in. More importantly,
tensix extension is not giving that much of perf benefit (at most 20-30%
in 2d cases). Until we find a way to double the perf with tensix
extension, this remains default disabled.

### What's changed
Only use row dispatch in BH when both tensix extension and fabric config
are specified by user, otherwise force to col dispatch.

### Checklist
- [ ] [All post commit]
https://github.com/tenstorrent/tt-metal/actions/runs/19144304535
- [ ] BH pc
https://github.com/tenstorrent/tt-metal/actions/runs/19145904273
- [ ] [Blackhole multi-card]
https://github.com/tenstorrent/tt-metal/actions/runs/19144256652
- [ ] BH demo
https://github.com/tenstorrent/tt-metal/actions/runs/19145922904
- [ ] (For runtime and ops writers) [T3000 nightly tests]
https://github.com/tenstorrent/tt-metal/actions/runs/19144500362