[GRID SAMPLE] Implementation of the nearest mode grid sample op (#32167)
### Problem description
Nearest mode of the grid sample op was missing support in the grid
sample op base.

### What's changed
Nearest mode of the op consists of 2 writer kernels and 0 compute
kernels. Its function is to calculate the nearest neighbor coordinates
beforehand through prepare grid op and read the corresponding input
sticks into the output cb which is sharded. Output of this mode is
always sharded. Input tensor is always L1 interleaved. The grid tensor
can be sharded or DRAM interleaved. Both of these are supported by the
same writer kernel. When the grid is sharded, output shard taketakes the
grid memory config. If the grid is not sharded, output config is either
calculated or the one asked for by a user is respected. Depending on the
align corners true which is a flag of the operation which indicates
whether the pixel point is in its center or corner the normalization of
the nearest neighbor changes a bit in calculation. Nearest mode of the
grid sample supports both align corners true and false. Currently there
is a started wip regarding the nearest mode supporting non precomputed
grid, but due to the bug that needs fixing it is guarded for in validate
method.

### Checklist
- [x] [All post-commit
tests](https://github.com/tenstorrent/tt-metal/actions/runs/19233537466)
- [x] [(Blackhole) Blackhole nightly
tests](https://github.com/tenstorrent/tt-metal/actions/runs/19233567583)
- [ ] [(Single-card) Device perf
regressions](https://github.com/tenstorrent/tt-metal/actions/runs/19233465725)
- [x] [(Single-card) Frequent model and ttnn
tests](https://github.com/tenstorrent/tt-metal/actions/runs/19233482838)
- [x] [Nightly tt-metal L2
tests](https://github.com/tenstorrent/tt-metal/actions/runs/19299078849)