Optimize Welfords LLK to hold intermediates for multiple groups at once (#797)
### Ticket
Issue: https://github.com/tenstorrent/tt-metal/issues/29706
This change accompanies the metal PR
https://github.com/tenstorrent/tt-metal/pull/31732
Please see the description of that PR as well for full context.

### Problem description
Current welford implementation updates the mean and variance as it scans
an entire tile (or a subset of rows in the tile). When applied to
groupnorm, there are instances when parts of a single tile belong to
multiple groups. To save on data movement, it is prudent to calculate
the online mean and var for each of these groups when we have the tile
fetched.

### What's changed
This PR does the following:
1. Adds functions to maintain up to 16 mean and vars simultaneously in a
tile. Earlier, the 32 mean values for a tile were stored in the first 4
rows of a tile in destination registers. With this change, the "rows"
are picked according to `group_id`
2. Reduce welford algorithm's instructions from 9 to 8 by using MOV
instructions. Use the entire replay buffer by directly operating out of
LREG0, LREG1 etc. and placing these in the replay buffer.
3. Separate out load, compute and store to give better control to kernel
writers and allow for optimizations.
4. Write down docstrings for all functions.
5. Use `TT_` and `TTI_` instructions and templates to move as many
things as possible to compile time calculations.

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update

### Checklist
<!-- These are required steps and need to be run from tt-metal
repository's Actions. Use links below and replace them with your run -->
- [ ] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes
- [ ] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml)
CI passes (if applicable)