Add SFPU kernel for ADD top row of two tiles (#807)
### Ticket
<!-- Link to Github Issue -->

https://github.com/tenstorrent/tt-metal/issues/29535#issuecomment-3479884917
### Problem description
<!-- Provide context for the problem. -->
The column reduction kernels store their results in the top row of each
tile. If you want to add two reduced tiles together, you would be
summing all the data elements in the tiles, even though only the top row
matters. This is computationally heavy and inefficient.
### What's changed
<!-- Describe the approach used to solve the problem.
Summarize the changes made and its impact. -->
Once the desired tiles have been reduced, if you want to accumulate
their sums, this kernel computes the sum using only the first row of
each tile. This approach greatly reduces the number of operations and
cycles required.
### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update