Use typed activation function param for matmul inside conv2d (#32367)
This change updates the handling of activation functions for matmul
operations inside conv2d by passing the typed UnaryOpWithParam directly,
instead of converting it to a string and passing that string to
ttnn::linear.

Recent updates to the matmul interfaces added native support for
UnaryOpWithParam to specify activation functions. Leveraging this
eliminates the need for the previous mapping from UnaryOpWithParam to
string representations.

The only caveat is that when matmul is sharded, activations must be
passed through the matmul program configuration. This approach is
necessary to ensure the activation function fusion is correctly applied.

Note that in this scenario, activation functions need to be set to
nullopt on the ttnn::linear interfaces, because even setting the same
activation directly leads to a different code path where fusion might
not occur.

Nightly tt-metal L2 tests
https://github.com/tenstorrent/tt-metal/actions/runs/19327013226
(Single-card) Frequent model and ttnn tests
https://github.com/tenstorrent/tt-metal/actions/runs/19326990874

Co-authored-by: Pavle Josipovic <pjosipovic@tenstorrent.com>