Emulate collectives in SingleHostContext
Message passing operations exposed by the DistributedContext are not
implemented in the SingleHostContext specialisation, or rather they
simply throw. Anywhere those ops are called we'll need to guarantee they
are not called when running with a SingleHostContext. This can't be
helped for send/recv, but collectives can be emulated (they are pretty
much nop in all cases). So, implement collective ops for SingleHostContext.

Signed-off-by: Piotr Stankiewicz <pstankiewicz@tenstorrent.com>