Fixing page buffer overflow in ReadFromDeviceInterleavedContiguous (#32227)
### Ticket
[#31998
](https://github.com/tenstorrent/tt-metal/issues/31998)
### Problem description
In `ReadFromDeviceInterleavedContiguous`, the `page` buffer to hold data
read-in from the device memory is hitting a buffer overflow. This is
because the buffer assumes a 32-bit datatype being read-in from the
device, and thus incorrectly uses `sizeof(uint32_t)` to determine how
many items (indices) to allocate inside `page.data()`. This is showing
up as `glibc` errors when running in slow dispatch mode on tests for the
`slice_rm_multi_core_stride` strided slice kernel, as insufficient space
is allocated on the buffer compared to the data being read-in.

### What's changed
Changed the `page` from a vector of `uint32_t` to a byte vector. This
removes dependence on the datatype stored on the device side, as we only
care about reading bytes.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/19306952502)
CI passes
- [x] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/19255237259)
CI with demo tests passes (if applicable)
- [x] New/Existing tests provide coverage for changes