[skip ci] Adding Auto Triage Workflow To CI (#32981)
### Ticket
[Link to Github
Issue](https://github.com/tenstorrent/metal-internal-workflows/issues/738)

### Problem description
This adds a CI workflow to run the action that will automatically triage
pipelines.

### What's changed
OpenCode CLI has been used along with a paid openAI API key to run
agentic models across the codebase. This allows gpt-codex or
gpt-codex-mini to download error logs and commit information to
determine which commits have caused pipelines to break. The main code to
execute this functionality is in the metal-infra-actions repository:
https://github.com/tenstorrent/metal-infra-actions

the purpose of this PR is to add the workflow to main so I can start
testing its functionality in CI. I might also make a local version as
well in a future PR.