[skip ci] Remove IWYU for now due to build failure (#32989)
### Ticket
N/A

### Problem description
IWYU is failing to build for what looks to be an LLVM version mismatch.
Something's clearly not right, but also it's not clear that anybody is
even using IWYU. So just remove it for now and we can look deeper if
it's actually being used by someone.

### What's changed
Removed IWYU.