Initial implementation of L1 memory accessor (#32856)
### Ticket
https://github.com/tenstorrent/tt-metal/issues/29592

### Problem description
- Implement a safe L1 memory accessor

### What's changed
- Implement a CoreLocalMem class with Watcher assertions when the user
tries to access data it's pointing to. Note, these Watcher asserts will
be upgraded to a Watcher Sanitize check in a future PR for better error
messages.
- User can still access to memory without the safety checks using
`get_unsafe_ptr()`
- Add unit tests to verify CoreLocalMem read, write, pointer math, and
it works as a source with the NoC 2.0 API, verify that there is no perf
impact when using this wrapper class compared to using raw pointers

### Checklist
All Post Commit
https://github.com/tenstorrent/tt-metal/actions/runs/19583021822
Blackhole Post Commit
https://github.com/tenstorrent/tt-metal/actions/runs/19583023364