Move fabric APIs into experimental folder (#32957)
### Ticket
No ticket

### Problem description
Runtime is stablizing our public host API. We aim to start lifecycling
changes in the `api/tt-metalium` folder.
Fabric has been identified as a project that is still in the active
development stage. It belongs in the `experimental` directory (and
namespace), where it can evolve in FAFO-style until it's stable enough
to join the core API.

This PR is the **first step** towards moving fabric into experimental.
It does the file movement only, which allows `tt-fabric` to bypass the
stability-oriented code review we are imposing on `api/metalium`.
Additional PRs are required to completely move fabric into experimental.

### What's changed

This PR is the step 1 of the movement, this does not update the
namespace instead simply performs file-movements & include updates.

Moved six headers into `api/tt-metalium/experimental/fabric`:
- control_plane.hpp
- edm_fabric_counters.hpp
- fabric_edm_types.hpp
- fabric_types.hpp
- fabric.hpp
- mesh_graph_descriptor.hpp
- mesh_graph.hpp
- routing_table_generator.hpp

- [x] PR is buildable.
- [x] Updated the codeowner file to better reflect ownership.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/19650200805)