Remove tt-metalium API directory from device side includes (#33086)
### Ticket
Partially related: #31083 , #20036

### Problem description
Currently, device is able to include symbols from the `tt-metalium`
folder which is intended for host APIs.

### What's changed
This PR does not address the issue but removes the direct directory
includes for `tt-metalium`, this allows includes of host side APIs to be
more descriptive and grepable so it's easier to cleanup in the future.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/19650310428)