Move return_output_dim to __call__ of TtConv2d
Since Conv2dConfiguration is inferred, and decision whether to return output dimensions is a user decision,
return_output_dim is given as arg when TtConv2d is called. That way no code that previously did not
need output dimensions is affected. In places where output dimensions were hardcoded now the arg is passed.