#33351: conv2d output grid override (#33352)
#33351

### What's changed
Added a parameter `override_output_sharding_config` to `Conv2dConfig` to
enable overriding the grid of the conv output tensor. This allows user
to control/reduce the grid on which matmul will be executed.

The reduction of the grid is done to in order to avoid di/dt issues on
WH HW.

Constraints of the feature are the following:

- not supported in the DRAM slicing case
- only supported for BS convs
- NHW num cores must match between input and output tensors

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/19761494649)
CI passes
- [x] [Nightly tt-metal L2
tests](https://github.com/tenstorrent/tt-metal/actions/runs/19766004832)
CI passes