Strided all gather to pair with experimental matmul (#32877)
### Problem description
The current all_gather is not efficient when fusing with matmul. In
particular, all_gather waits for the entire slice to show up before
signaling to matmul to begin processing the slice. Worse yet, once
matmul processes the first row of K blocks associated with an all-gather
device slice, it stalls waiting for the next full next slice to show up
before it can proceed, instead of processing the rest of the slice it
just received. The alternative would be to spill a partial row of the
matmul results, but this would take matmul efficiency.

### What's changed
As a result, we created strided_all_gather_async. This changes the
schedule of the all gather. It sends a row of K_blocks (same as before),
but instead of proceeding to the next row, it strides to the row that
the next matmul core would be using, and sends that. After sending one
set of "rows" of K_blocks that would allow all matmul cores to continue
processing, it signals. Moreover, it waits to receive the next set of
K_blocks from the same rows, not moving on to the next row until this
full row is done. This allows the matmul to proceed in it's natural
order without unnecessary spills (finishing one entire row of K_blocks
and writing to output once before moving on), allowing for better
overlap between all gather and matmul.

To go with the strided_all_gather_async, there is a
strided_all_gather_minimal_matmul_async that fuses with the experimental
matmul.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/19811234658)
- [x] [T3000 unit
tests](https://github.com/tenstorrent/tt-metal/actions/runs/19641945213)
- [x] [Galaxy units
tests](https://github.com/tenstorrent/tt-metal/actions/runs/19641960208)