Fix Distributed RMSNorm hang caused by CB sizing (#33254)
### Problem description
Certain shapes led to hangs in
`ttnn.experimental.wan_fused_rmsnorm_post_allgather` due to incorrect
sizing of the weight CB.

### What's changed
Ensure that weight CB is sized as a multiple of the DST granularity.
This is in alignment with how the weight CB is produced by reader and
consumed by compute.

### Checklist
- [x] L2 nightly
https://github.com/tenstorrent/tt-metal/actions/runs/19707831573