sfpu reduce subblock42 (#902)
### Ticket
[metal issue](https://github.com/tenstorrent/tt-metal/issues/26699)

### Problem description
On WH cards we need subblock to be of size 4x2 in order to get max
matmul perf. This change is also implemented for initial BH perf
measurements. In future BH version could be changed

### What's changed
Added _custom reduce sfpu files that contain code for this special case
ruduce, so they don't get confused with existing sfpu reduce max that
works on any subblock size

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update

---------

Co-authored-by: ldjurovicTT <ldjurovic@tenstoreent.com>