Decouple SDXL and SD 1.4 device perf tests (#33637)
### Problem description
SDXL and SD 1.4 perf tests are coupled together within `stable_diffusion
tests` step of the `N300 WH B0 Set 1 device perf` job in `(Single-card)
Device perf regressions`.

This leads to confusion as the mentioned models are not the same, they
have different implementations and are owned by separate teams. In
addition to this SDXL is a model in production and needs to be closely
monitored in the CI. This proves difficult if its counterpart (tested
before it in the CI) fails and the test set errors out before testing
the second model.

### What's changed
Tests for the mentioned models are split into separate steps in the job.

### Checklist
- [x] [(Single-card) Device perf
regressions](https://github.com/tenstorrent/tt-metal/actions/runs/19886404904/job/56994919440)
CI passes