fix to layout regression (#33665)
### Ticket
https://github.com/tenstorrent/tt-metal/issues/33620

### Problem description
To layout was having a segmentation fault because some code that was not
updated would fail if the number of cores required equaled a multiple of
the width of the device grid plus one. This was a legacy piece of code
when the core grid was assumed to be rectangular always which prevented
this condition from ever happening and it was necessary if one assumes
the old way of segmenting core grids before my PR changed that.

Due to the subtle nature of the error it was not captured in any of our
standard to layout or tilize tests but it did show up in a model case.
Testing for this condition was added to the test suite

### What's changed
I deleted the convoluted core logic and simply added a line to add the
cliff core to the list of all cores. With the modifications I made to
delete the rectangular requirements this code is no longer necessary.
This patches the error and the simplified logic will be less error prone
for other edge cases.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes
https://github.com/tenstorrent/tt-metal/actions/runs/19899131428
- [ ] [Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml)
CI with demo tests passes (if applicable)
https://github.com/tenstorrent/tt-metal/actions/runs/19899869945
- [ ] T3k Model Perf
https://github.com/tenstorrent/tt-metal/actions/runs/19899908208
- [x] T3k Frequent
https://github.com/tenstorrent/tt-metal/actions/runs/19899934530