Enabling tt-triage in APC (#33798)
We decided to proceed with enabling tt-triage in CI in stages.
First stage is enabling tt-triage in APC and Merge gate, so we can see
how it performs.
Second stage is enabling it in all other workflows and asking people to
report to us when they see failure that shouldn't be there (timeout was
too short).

Here is [APC
run](https://github.com/tenstorrent/tt-metal/actions/runs/19932245278)
for this PR. Note that failed t3000-apc-fast-tests were not touched by
this change.