[skip ci] Refactor wheel building CI job (#33530)
### Ticket
N/A

### Problem description
We wish to build wheels for many versions of Python. Currently the wheel
build is tied to Ubuntu builds. Since the move to manylinux, this no
longer is necessary. For building against a sweep of Python versions,
this hinders us.

### What's changed
Refactor the workflow job so we can build the wheel against arbitrary
Python versions.

Workflows with forced "run everything":
PR Gate:
https://github.com/tenstorrent/tt-metal/actions/runs/19839909326
Merge Gate:
https://github.com/tenstorrent/tt-metal/actions/runs/19840153052