make t3k fabric BW test tolerances bigger to only catch large violations (#33868)
This is in response to the current approach which is not working well
for the team. Even very minor performance changes can throw tests out of
tolerance when those tests are run across multiple machines. This will
make it so failures are likely only to happen when big performance
regressions are identified. Performance regressions in general will
instead be monitored on superset.