Allow creating MeshDevice spanning subset of ranks (whole fabric init)
When initialising a MeshDevice which does not span the full set of ranks
in a distributed set up, we still want to allow the full fabric init to
run. So, make the "root" MeshDevice creation process more like sub-mesh
creation. I.e. always create ScopedDevices for the full system mesh
shape, but limit the shape of the created MeshDevice to the requested
one.

Signed-off-by: Piotr Stankiewicz <pstankiewicz@tenstorrent.com>