disable dcache usage in fabric on BH (#33509)
The riscv (erisc) datacache (along with it's invalidation mechanism)
isn't very performance compatible with typical memory access patterns.

The fabric makes regular use of the following pattern:
1) Status check some credit register (typically an overlay register)
2) Immediately fetch the latest data (e.g. packet header) into data
cache
- on BH, this is only possible by calling the fence operation, which
will materialize as a completion of all outstanding reads/writes.

The above is not compatible to achieving best performance with data
cache enabled. To better leverage data-cache, alternative
implementations would be required - those that don't require such
frequent invalidation. That is separate work and in the mean-time, the
data-cache is disabled, leading to a nearly 10% speedup on BH fabric
workloads.

Closes #33358 

### Checklist
- [x] T3K Fast:
https://github.com/tenstorrent/tt-metal/actions/runs/19942430099
- [x] BH Multi-card Unit Test:
https://github.com/tenstorrent/tt-metal/actions/runs/19992677863