Revert removal of `get_pcie_alignment` in hal (#33981)
### Ticket
Closes https://github.com/tenstorrent/tt-mlir/issues/6220

### Problem description
`get_pcie_alignment` was removed in #32115 as it was thought that no one
is dependent on it.
However this turned out not true as mlir is an active user of the
function and the removal has affected their uplift.

### What's changed
Revert removal of `get_pcie_alignment` function in Runtime host api.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/20021231310)