[skip ci] ci: provide more information when doing the llk uplift (added issue information) (#34224)
### Ticket
None

### Problem description
At present, table that gets generated from LLK commits does not contain
information on the issue number. Since GH links these with PRs, it'd be
a nice thing to have, as that way we could track on the issue itself if
it got all the way to metal or it's still stuck somewhere in between.

### What's changed
With this change, we now add information about the issue directly in the
changelog table. That way GH does the tracking automatically and links
the change with metal uplift PR.