Integrate watcher sanitize into safe L1 accessor (#34275)
### Ticket
https://github.com/tenstorrent/tt-metal/issues/32949

### Problem description
Safe L1 access functions were using asserts which do not provide that
much information.
They did not take into account if the kernel was on Ethernet or Tensix

### What's changed
Update them to use the L1 sanitize functions which will show a
informative error message when an invalid L1 address is accessed. This
watcher check compiles with the correct max L1 address depending on if
it's Eth or Tensix

### Checklist
All Post Commit
https://github.com/tenstorrent/tt-metal/actions/runs/20177145271
Blackhole Post Commit
https://github.com/tenstorrent/tt-metal/actions/runs/20177143525